test/nginx: extract request() and add testing
