+++ test/nginx/src/mocha/nginx.spec.js
@@ -1,5 +1,4 @@
 const tls = require('node:tls');
-const { Readable } = require('stream');
 
 const {
   assert,
@@ -7,6 +6,7 @@ const {
   requestSentryMock,
   resetSentryMock,
 } = require('../lib');
+const request = require('./request');
 
 const none = `'none'`;
 const reportSample = `'report-sample'`;
@@ -1175,61 +1175,6 @@ function resetBackendMock() {
 async function resetMock(port) {
   const res = await request(`http://localhost:${port}/reset`);
   assert.isTrue(res.ok);
-}
-
-// Similar to fetch() but:
-//
-// 1. do not follow redirects
-// 2. allow overriding of fetch's "forbidden" headers: https://developer.mozilla.org/en-US/docs/Glossary/Forbidden_header_name
-function request(url, { body, ...options }={}) {
-  if(!options.headers) options.headers = {};
-  if(!options.headers.host) options.headers.host = 'odk-nginx.example.test';
-
-  return new Promise((resolve, reject) => {
-    try {
-      const req = getProtocolImplFrom(url).request(url, options, res => {
-        res.on('error', reject);
-
-        const body = new Readable({ read:() => {} });
-        res.on('error', err => body.destroy(err));
-        res.on('data', data => body.push(data));
-        res.on('end', () => body.push(null));
-
-        const text = () => new Promise((resolve, reject) => {
-          const chunks = [];
-          body.on('error', reject);
-          body.on('data', data => chunks.push(data));
-          body.on('end', () => resolve(Buffer.concat(chunks).toString('utf8')));
-        });
-
-        const status = res.statusCode;
-
-        resolve({
-          status,
-          ok: status >= 200 && status < 300,
-          statusText: res.statusText,
-          body,
-          text,
-          json: async () => JSON.parse(await text()),
-          headers: new Headers(res.headers),
-        });
-      });
-      req.on('error', reject);
-      if(body !== undefined) req.write(body);
-      req.end();
-    } catch(err) {
-      reject(err);
-    }
-  });
-}
-
-function getProtocolImplFrom(url) {
-  const { protocol } = new URL(url);
-  switch(protocol) {
-    case 'http:':  return require('node:http');
-    case 'https:': return require('node:https');
-    default: throw new Error(`Unsupported protocol: ${protocol}`);
-  }
 }
 
 function assertCacheStrategyApplied(res, expectedCacheStrategy) {
